Fix tests to not rely on Chrome being on the underlying machine
